zabbixui: Revamp container for Zabbix 5.0
As we no longer want to use CentOS as a base, I wanted to normalize the
container with our other deployed containers, bringing the differences down
even more.

This container still differs from our instance container, by the version and
configuration files.

However, the deployments are now more similar, and we shall work to make them
even more so in the future.

The configuration of Nginx and php-fpm has been re-syncronized with upstream
Debian Zabbix UI packages.

Issue: ModioAB/agile#1862