Add walg as a container as well
This uses an non-integrated container with Walg instead of built in functions
in spilo.

The container is based on the bitnami one mostly as a test, as some testing in
devleopment showed that it worked for this use-case.

The changes to the container are only to make sure that the user matches that
in the patroni/spilo image to avoid issues with nfs uids.